set unets as classes (wip)
